requirements.txt
setuptools>=42
wheel
paho-mqtt==1.6.1
bluepy==1.3.0
hatchrestbluepy==1.1.1
